+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 ### Changed
+* Versioning between client and server is no longer tightly coupled.  Servers can warn or block older out-of-date clients.
 
 ### Fixed